+++ .eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   extends: [ '@cenk1cenk2/eslint-config/typescript' ],
   rules: {
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
     'import/order': [
       'error',
       {