+++ tests/concurrency.e2e-spec.ts
@@ -4,7 +4,6 @@ import { ListrTask } from '@interfaces/listr.interface'
 import { Listr } from '@root/index'
 
 describe('concurrent execution', () => {
-
   let tasks: ListrTask<any, any>[]
   let log: jest.SpyInstance<void, string[][]>
 
@@ -12,37 +11,34 @@ describe('concurrent execution', () => {
     log = jest.spyOn(console, 'log').mockImplementation()
 
     tasks = [
-
       {
         title: '1',
         task: async (): Promise<void> => {
-          await delay(20)
+          await delay(28)
         }
       },
 
       {
         title: '2',
         task: async (): Promise<void> => {
-          await delay(15)
+          await delay(16)
         }
       },
 
       {
         title: '3',
         task: async (): Promise<void> => {
-          await delay(10)
+          await delay(9)
         }
       },
 
       {
         title: '4',
         task: async (): Promise<void> => {
-          await delay(1)
+          await delay(0)
         }
       }
-
     ]
-
   })
 
   afterEach(async () => {
@@ -54,66 +50,66 @@ describe('concurrent execution', () => {
 
     expect(log).toBeCalledTimes(8)
     expect(log.mock.calls).toMatchInlineSnapshot(`
-    Array [
-      Array [
-        "[STARTED] 1",
-      ],
-      Array [
-        "[STARTED] 2",
-      ],
-      Array [
-        "[STARTED] 3",
-      ],
-      Array [
-        "[STARTED] 4",
-      ],
-      Array [
-        "[SUCCESS] 4",
-      ],
-      Array [
-        "[SUCCESS] 3",
-      ],
-      Array [
-        "[SUCCESS] 2",
-      ],
-      Array [
-        "[SUCCESS] 1",
-      ],
-    ]
-      `)
+          Array [
+            Array [
+              "[STARTED] 1",
+            ],
+            Array [
+              "[STARTED] 2",
+            ],
+            Array [
+              "[STARTED] 3",
+            ],
+            Array [
+              "[STARTED] 4",
+            ],
+            Array [
+              "[SUCCESS] 4",
+            ],
+            Array [
+              "[SUCCESS] 3",
+            ],
+            Array [
+              "[SUCCESS] 2",
+            ],
+            Array [
+              "[SUCCESS] 1",
+            ],
+          ]
+          `)
   })
 
   it('should limit the concurrency', async () => {
-    await new Listr(tasks, { concurrent: 2, renderer: 'verbose' }).run()
+    await new Listr(tasks, { concurrent: 3, renderer: 'verbose' }).run()
 
     expect(log).toBeCalledTimes(8)
     expect(log.mock.calls).toMatchInlineSnapshot(`
-    Array [
       Array [
-        "[STARTED] 1",
-      ],
-      Array [
-        "[STARTED] 2",
-      ],
-      Array [
-        "[SUCCESS] 2",
-      ],
-      Array [
-        "[STARTED] 3",
-      ],
-      Array [
-        "[SUCCESS] 1",
-      ],
-      Array [
-        "[STARTED] 4",
-      ],
-      Array [
-        "[SUCCESS] 4",
-      ],
-      Array [
-        "[SUCCESS] 3",
-      ],
-    ]
+        Array [
+          "[STARTED] 1",
+        ],
+        Array [
+          "[STARTED] 2",
+        ],
+        Array [
+          "[STARTED] 3",
+        ],
+        Array [
+          "[SUCCESS] 3",
+        ],
+        Array [
+          "[STARTED] 4",
+        ],
+        Array [
+          "[SUCCESS] 4",
+        ],
+        Array [
+          "[SUCCESS] 2",
+        ],
+        Array [
+          "[SUCCESS] 1",
+        ],
+      ]
     `)
   })
 
@@ -122,33 +118,32 @@ describe('concurrent execution', () => {
 
     expect(log).toBeCalledTimes(8)
     expect(log.mock.calls).toMatchInlineSnapshot(`
-    Array [
-      Array [
-        "[STARTED] 1",
-      ],
-      Array [
-        "[SUCCESS] 1",
-      ],
-      Array [
-        "[STARTED] 2",
-      ],
-      Array [
-        "[SUCCESS] 2",
-      ],
-      Array [
-        "[STARTED] 3",
-      ],
-      Array [
-        "[SUCCESS] 3",
-      ],
-      Array [
-        "[STARTED] 4",
-      ],
-      Array [
-        "[SUCCESS] 4",
-      ],
-    ]
-    `)
+          Array [
+            Array [
+              "[STARTED] 1",
+            ],
+            Array [
+              "[SUCCESS] 1",
+            ],
+            Array [
+              "[STARTED] 2",
+            ],
+            Array [
+              "[SUCCESS] 2",
+            ],
+            Array [
+              "[STARTED] 3",
+            ],
+            Array [
+              "[SUCCESS] 3",
+            ],
+            Array [
+              "[STARTED] 4",
+            ],
+            Array [
+              "[SUCCESS] 4",
+            ],
+          ]
+        `)
   })
-
 })